refactor(components): refactor icon (#3528)
